feat: add current route  highlighting
