+++ package.json
@@ -38,6 +38,7 @@
     "axios": "^0.21.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^5.2.0",
     "reinstall": "^2.0.0"
   }
 }
