+++ src/components/about/about.scss
@@ -1,0 +1,6 @@
+// h3 {
+//   transition: all 1s;
+//   &:hover {
+//     color: lime;
+//   }
+// }
