+++ src/components/header/header.scss
@@ -1,0 +1,28 @@
+.header {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+}
+.nav {
+  ul {
+    display: flex;
+    gap: 30px;
+    margin-bottom: 10%;
+    list-style: none;
+    a {
+      color: green;
+      text-decoration: none;
+      transition: all 0.5s;
+      &:hover {
+        // background: linear-gradient(lightskyblue, lemonchiffon);
+        background: lightblue;
+      }
+    }
+  }
+}
+.active {
+  // background: linear-gradient(lemonchiffon, lightskyblue);
+  background: lemonchiffon;
+  box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.15);
+}
