+++ src/styles.scss
@@ -8,6 +8,9 @@ html {
   font-size: 20px;
 }
 #rootNode {
+  display: flex;
+  flex-direction: column;
+  gap: 30px;
   // height: 100vh;
   width: 100vw;
   color: lightblue;
